refactor: 검색결과 페이지 컴포넌트 분리
